finished making ChatRoom reusable
