+++ client/src/components/exercise/lab8/components/ChatRoom.js
@@ -1,55 +1,10 @@
 import React, { useState, useEffect } from "react";
 import ChatMessage from "./ChatMessage";
-import { CHAT_MESSAGES } from "../../../../constants/lab8";
 
 import PropTypes from "prop-types";
 
-// Select random messages to display in the chat room.
-function selectMessages() {
-  const totalMessages = Math.floor(Math.random() * 3) + 6; // Random number between 4 and 6.
-  const halfMessages = Math.floor(totalMessages / 2);
-
-  // Process all the messages into buckets based on their type
-  const incorrectKeep = CHAT_MESSAGES.recommend_keep.filter(
-    (msg) => !msg.ai_correct
-  );
-  const incorrectRemove = CHAT_MESSAGES.recommend_remove.filter(
-    (msg) => !msg.ai_correct
-  );
-
-  const correctKeep = CHAT_MESSAGES.recommend_keep.filter(
-    (msg) => msg.ai_correct
-  );
-  const correctRemove = CHAT_MESSAGES.recommend_remove.filter(
-    (msg) => msg.ai_correct
-  );
-
-  // Select random messages from each bucket to display in the chat room.
-  const selectedKeep = [
-    ...shuffleArray(incorrectKeep).slice(0, 1),
-    ...shuffleArray(correctKeep).slice(0, halfMessages - 1),
-  ];
-
-  const selectedRemove = [
-    ...shuffleArray(incorrectRemove).slice(0, 1),
-    ...shuffleArray(correctRemove).slice(0, halfMessages - 1),
-  ];
-
-  // Shuffle the messages and return them.
-  return shuffleArray([...selectedKeep, ...selectedRemove]);
-}
-
-function shuffleArray(array) {
-  const newArray = [...array];
-  for (let i = newArray.length - 1; i > 0; i--) {
-    const j = Math.floor(Math.random() * (i + 1));
-    [newArray[i], newArray[j]] = [newArray[j], newArray[i]];
-  }
-  return newArray;
-}
-
 const ChatRoom = (props) => {
-  const { moderationCompleteCallback } = props;
+  const { moderationCompleteCallback, selectMessages } = props;
 
   const [messages] = useState(selectMessages());
   const [currentIndex, setCurrentIndex] = useState(0);
@@ -113,6 +68,7 @@ const ChatRoom = (props) => {
 
 ChatRoom.propTypes = {
   moderationCompleteCallback: PropTypes.func.isRequired,
+  selectMessages: PropTypes.func.isRequired,
 };
 
 export default ChatRoom;
